backend/requirements.txt
# Django and Django REST Framework
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1

# Authentication
djangorestframework-simplejwt==5.3.0

# Background Tasks
celery==5.3.6
redis==5.0.1

# Environment variables
python-decouple==3.8

# YAML support for service configs
PyYAML==6.0.1
gunicorn==21.2.0
